+++ lib/main.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
-import 'dart:ffi';
 
 import 'package:asset_cache/asset_cache.dart';
 import 'package:ffi/ffi.dart';
